Parse row hints for LSJ
This commit adds support to parse row hints for LSJ.
A testcase was added such that uses LSJ to derive stats from group.
This is done by disabling xforms that create InnerJoin.
This is because statistics for the group are derived from the
expression that has highest stats promise within the group.
LSJ has a medium stat promise whereas IJ often has a
high stats promise.